[SourceKit] Share common object printer code
This is an attempt to clean up code duplication around printing SourceKit
request and response objects.
